+++ ride-sharing/src/main/java/com/example/geektrust/repository/RiderRepo.java
@@ -1,6 +1,5 @@
 package com.example.geektrust.repository;
 
-import com.example.geektrust.model.Driver;
 import com.example.geektrust.model.Rider;
 
 import java.util.HashMap;
@@ -8,7 +7,7 @@ import java.util.Map;
 
 public class RiderRepo {
     private Map<String, Rider> riderMap;
-    public boolean addRider(String id, Integer x, Integer y){
+    public boolean addRider(String id, Double x, Double y){
         if(riderMap.containsKey(id))
             return false;
         riderMap.put(id,new Rider(id,x,y));
